+++ src/api/components/pacientes/controller.ts
@@ -1,12 +1,13 @@
 import {Request, Response} from 'express'
 import { PatientService } from './service'
 import logger from '../../../utils/logger'
+import { createPatientSchema, identificacionSchema, updatePatientSchema } from './validations/patient.validation'
 
 export interface PatientController {
     getAllPatients(req: Request, res: Response): Promise<void>
     createPatient(req: Request, res: Response): Promise<void>
     getPatientById(req: Request, res: Response): Promise<void>
-    getPatientByCedula(req: Request, res: Response): Promise<void>
+    getPatientByIdentificacion(req: Request, res: Response): Promise<void>
     updatePatientById(req: Request, res: Response): Promise<void>
     deletePatientById(req: Request, res: Response): Promise<void>
 }
@@ -21,8 +22,8 @@ export class PatientControllerImpl implements PatientController {
     
     public async getAllPatients(req: Request, res: Response): Promise<void> {
         await this.patientService.getAllPatients()
-        .then((getAllPatients)=>{
-            res.status(200).json(getAllPatients);
+        .then((allPatients)=>{
+            res.status(200).json(allPatients);
         },
         (error) => {
             res.status(400).json({message: `${error.message}`}) 
@@ -30,8 +31,12 @@ export class PatientControllerImpl implements PatientController {
     }
 
     public async createPatient(req: Request, res: Response): Promise<void> {
-        const patientReq = req.body
-        await this.patientService.createPatient(patientReq)
+        const {error, value} = createPatientSchema.validate(req.body);
+        if (error){
+            res.status(400).json({message: `${error.details[0].message}`});
+            return;
+        }
+        await this.patientService.createPatient(value)
         .then((patient)=>{
             logger.info(`New patient created succesfully: ${JSON.stringify(patient)}`)
             res.status(201).json(patient)
@@ -42,7 +47,11 @@ export class PatientControllerImpl implements PatientController {
     }
 
     public async getPatientById(req: Request, res: Response):Promise<void> {
-        const id = req.params.id
+        const id = parseInt(req.params.id)
+        if(isNaN(id)){
+            res.status(400).json({message: `Error: ID must be a number`});
+            return
+        }
         await this.patientService.getPatientById(id)
         .then((patient)=>{
             res.status(200).json(patient)
@@ -52,10 +61,13 @@ export class PatientControllerImpl implements PatientController {
         })
     }
 
-    public async getPatientByCedula(req: Request, res: Response):Promise<void> {
-        const patientReq = req.body
-        const identificacion = patientReq.identificacion
-        await this.patientService.getPatientByCedula(identificacion)
+    public async getPatientByIdentificacion(req: Request, res: Response):Promise<void> {
+        const {error, value} = identificacionSchema.validate(req.body);
+        if (error){
+            res.status(400).json({message: `${error.details[0].message}`});
+            return;
+        }
+        await this.patientService.getPatientByIdentificacion(value.identificacion)
         .then((patient)=>{
             res.status(200).json(patient)
         },
@@ -65,9 +77,17 @@ export class PatientControllerImpl implements PatientController {
     }
 
     public async updatePatientById(req: Request, res: Response):Promise<void> {
-        const id = req.params.id
-        const patientReq = req.body; 
-        await this.patientService.updatePatientById(id, patientReq)
+        const id = parseInt(req.params.id);
+        if(isNaN(id)){
+            res.status(400).json({message: `Error: ID must be a number`});
+            return
+        }
+        const {error, value} = updatePatientSchema.validate(req.body);
+        if (error){
+            res.status(400).json({message: `${error.details[0].message}`});
+            return;
+        }
+        await this.patientService.updatePatientById(id, value)
         .then((patient)=>{
             logger.info(`Patient updated succesfully: ${JSON.stringify(patient)}`)
             res.status(200).json(patient)
@@ -78,7 +98,11 @@ export class PatientControllerImpl implements PatientController {
     }
 
     public async deletePatientById(req: Request, res: Response):Promise<void> {
-        const id = req.params.id
+        const id = parseInt(req.params.id);
+        if(isNaN(id)){
+            res.status(400).json({message: `Error: ID must be a number`});
+            return
+        }
         await this.patientService.deletePatientById(id)
         .then((patient)=>{
             res.status(200).json({message: "Patient deleted succesfully"})
